refactor(IaaC) spit the stack to 4 part (front, backend, db, deployment)
